fix(wasm-fs): simplify error translation and reorganize polyfills
